Moving test op mode files to new directory.
